requirements.txt
# Core ML and Deep Learning
tensorflow>=2.13.0
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Data Processing and Scientific Computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
imbalanced-learn>=0.10.0

# Image Processing and Computer Vision
opencv-python>=4.8.0
pillow>=10.0.0
albumentations>=1.3.0
imageio>=2.33.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Development and Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipython>=8.15.0
notebook>=7.0.0

# Monitoring and Callbacks
tensorboard>=2.13.0
wandb>=0.15.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.66.0
pydantic>=2.0.0
loguru>=0.7.0

# API and Web Framework
flask>=2.3.0
fastapi>=0.103.0
uvicorn>=0.24.0
python-multipart>=0.0.6
requests>=2.31.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code Quality
black>=23.9.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.5.0
pre-commit>=3.5.0

# Optional: For edge computing
onnx>=1.14.0
onnxruntime>=1.16.0
